#12792. Remove legacy encryption mode based on RSA

+++ src/dbSchema.sql
@@ -32,7 +32,7 @@ CREATE TABLE history(idx int not null, chatid int64 not null, msgid int64 not nu
     userid int64, keyid int not null, type tinyint, updated smallint, ts int,
     is_encrypted tinyint, data blob, backrefid int64 not null, UNIQUE(chatid,msgid), UNIQUE(chatid,idx));
 
-CREATE TABLE sendkeys(chatid int64 not null, userid int64 not null, keyid int64 not null, key blob not null,
+CREATE TABLE sendkeys(chatid int64 not null, userid int64 not null, keyid int32 not null, key blob not null,
     ts int not null, UNIQUE(chatid, userid, keyid));
 
 CREATE TABLE node_history(idx int not null, chatid int64 not null, msgid int64 not null,
